xcel_itron2mqtt/requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
ifaddr==0.2.0
httpx
urllib3==1.26.14
zeroconf==0.47.3
pyyaml==6.0.1
tenacity==8.2.3
asyncio
asyncio-mqtt
